# src/main/apiServer/generated/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cherry Studio API",
    "version": "1.0.0",
    "description": "OpenAI-compatible API for Cherry Studio with additional Cherry-specific endpoints",
    "contact": {
      "name": "Cherry Studio",
      "url": "https://github.com/CherryHQ/cherry-studio"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use the API key from Cherry Studio settings"
      }
    },
    "schemas": {
...
      "CreateSessionMessageRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "description": "Message content"
          }
        },
        "required": ["content"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "Error message"
              },
              "type": {
                "type": "string",
                "description": "Error type"
              },
              "code": {
                "type": "string",
                "description": "Error code"
              }
            },
            "required": ["message", "type", "code"]
          }
        },
        "required": ["error"]
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/agents/{agentId}/sessions/{sessionId}/messages": {
      "post": {
        "summary": "Create a new message in a session",
        "tags": ["Messages"],
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Agent ID"
          },
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Session ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionMessageRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "description": "Message ID"
                    },
                    "session_id": {
                      "type": "string",
                      "description": "Session ID"
                    },
                    "role": {
                      "type": "string",
                      "enum": ["assistant", "user", "system", "tool"],
                      "description": "Message role"
                    },
                    "content": {
                      "type": "object",
                      "description": "Message content (AI SDK format)"
                    },
                    "agent_session_id": {
                      "type": "string",
                      "description": "Agent session ID for resuming"
                    },
                    "metadata": {
                      "type": "object",
                      "description": "Additional metadata"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agent or session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agents/{agentId}/sessions/{sessionId}/messages/{messageId}": {
      "delete": {
        "summary": "Delete a message from a session",
        "tags": ["Messages"],
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Agent ID"
          },
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Session ID"
          },
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Message ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Message deleted successfully"
          },
          "404": {
            "description": "Agent, session, or message not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/messages": {
      "post": {
        "summary": "Create message",
        "description": "Create a message response using Anthropic's API format",
        "tags": ["Messages"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["model", "max_tokens", "messages"],
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Model ID in format \"provider:model_id\"",
                    "example": "my-anthropic:claude-3-5-sonnet-20241022"
                  },
                  "max_tokens": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Maximum number of tokens to generate",
                    "example": 1024
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": ["user", "assistant"]
                        },
                        "content": {
                          "oneOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "array"
                            }
                          ]
                        }
                      }
                    }
                  },
                  "system": {
                    "type": "string",
                    "description": "System message"
                  },
                  "temperature": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1,
                    "description": "Sampling temperature"
                  },
                  "top_p": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1,
                    "description": "Nucleus sampling"
                  },
                  "top_k": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Top-k sampling"
                  },
                  "stream": {
                    "type": "boolean",
                    "description": "Whether to stream the response"
                  },
                  "tools": {
                    "type": "array",
                    "description": "Available tools for the model"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "example": "message"
                    },
                    "role": {
                      "type": "string",
                      "example": "assistant"
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "model": {
                      "type": "string"
                    },
                    "stop_reason": {
                      "type": "string"
                    },
                    "stop_sequence": {
                      "type": "string"
                    },
                    "usage": {
                      "type": "object",
                      "properties": {
                        "input_tokens": {
                          "type": "integer"
                        },
                        "output_tokens": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-sent events stream (when stream=true)"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "error"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/{provider_id}/v1/messages": {
      "post": {
        "summary": "Create message with provider in path",
        "description": "Create a message response using provider ID from URL path",
        "tags": ["Messages"],
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Provider ID (e.g., \"my-anthropic\")",
            "example": "my-anthropic"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["model", "max_tokens", "messages"],
                "properties": {
                  "model": {
                    "type": "string",
                    "description": "Model ID without provider prefix",
                    "example": "claude-3-5-sonnet-20241022"
                  },
                  "max_tokens": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Maximum number of tokens to generate",
                    "example": 1024
                  },
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "role": {
                          "type": "string",
                          "enum": ["user", "assistant"]
                        },
                        "content": {
                          "oneOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "array"
                            }
                          ]
                        }
                      }
                    }
                  },
                  "system": {
                    "type": "string",
                    "description": "System message"
                  },
                  "temperature": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1,
                    "description": "Sampling temperature"
                  },
                  "top_p": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1,
                    "description": "Nucleus sampling"
                  },
                  "top_k": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Top-k sampling"
                  },
                  "stream": {
                    "type": "boolean",
                    "description": "Whether to stream the response"
                  },
                  "tools": {
                    "type": "array",
                    "description": "Available tools for the model"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Message response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string",
                      "example": "message"
                    },
                    "role": {
                      "type": "string",
                      "example": "assistant"
                    },
                    "content": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "model": {
                      "type": "string"
                    },
                    "stop_reason": {
                      "type": "string"
                    },
                    "stop_sequence": {
                      "type": "string"
                    },
                    "usage": {
                      "type": "object",
                      "properties": {
                        "input_tokens": {
                          "type": "integer"
                        },
                        "output_tokens": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              },
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "Server-sent events stream (when stream=true)"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}